backend/requirements.txt
# Core ML Libraries
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
transformers==4.45.0
tokenizers>=0.19.0
datasets>=2.18.0
timm==1.0.11
xgboost==2.1.1
optuna==4.0.0

# FastAPI and Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.10
pydantic==2.9.0

# Data Processing
numpy>=2.1.0
pandas==2.2.3
Pillow>=10.4.0
opencv-python-headless>=4.10.0

# Vision and Image Processing
opencv-python-headless==4.12.0.88

# Scientific Computing
scikit-learn==1.5.2
scipy==1.14.1
matplotlib==3.9.2
seaborn==0.13.2
joblib==1.4.2

# Utilities
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5
wandb==0.18.0

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2